+++ EssentialApp/EssentialAppTests/FeedLoaderCacheDecoratorTests.swift
@@ -21,14 +21,14 @@ class FeedLoaderCacheDecoratorTests: XCTestCase {
 
 	func test_load_deliversFeedOnLoaderSuccess() {
 		let feed = uniqueFeed()
-		let loader = LoaderStub(result: .success(feed))
+		let loader = FeedLoaderStub(result: .success(feed))
 		let sut = FeedLoaderCacheDecorator(decoratee: loader)
 		
 		expect(sut, toCompleteWith: .success(feed))
 	}
 	
 	func test_load_deliversErrorOnLoaderFailure() {
-		let loader = LoaderStub(result: .failure(anyNSError()))
+		let loader = FeedLoaderStub(result: .failure(anyNSError()))
 		let sut = FeedLoaderCacheDecorator(decoratee: loader)
 		
 		expect(sut, toCompleteWith: .failure(anyNSError()))
@@ -60,18 +60,5 @@ class FeedLoaderCacheDecoratorTests: XCTestCase {
 	private func uniqueFeed() -> [FeedImage] {
 		return [FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())]
 	}
-
-	private class LoaderStub: FeedLoader {
-		private let result: FeedLoader.Result
-		
-		init(result: FeedLoader.Result) {
-			self.result = result
-		}
-
-		func load(completion: @escaping (FeedLoader.Result) -> Void) {
-			completion(result)
-		}
-	}
-
 	
 }
